refactor(server): extract nodes routes from server.ts (#3502 PR3 step 1)
Move the 20 inline /nodes* + /auto-favorite/status handlers, plus
/auto-ping/stop/:nodeNum, into routes/nodesRoutes.ts, mounted at '/'
alongside the existing '/'-mounted routers. Handlers moved verbatim;
only the leading apiRouter.* -> router.* rewrite and two relative
import-path fixes (nodeInfoCopyService, lowEntropyKeyService) were
made. server.ts drops to 1,357 lines.
